Switch the journeys walkTree function to use the new treeWalkers
