+++ src/flow/index.js
@@ -3,9 +3,25 @@ const Branch = require('./branch');
 const Stop = require('./stop');
 const journey = require('./journey');
 const RequestBoundJourney = require('./RequestBoundJourney');
+const {
+  ifCompleteThenContinue,
+  continueToNext,
+  stopHere,
+  validateThenStopHere
+} = require('./treeWalker');
 
 const goTo = step => new Redirector(step);
 const branch = (...redirectors) => new Branch(...redirectors);
 const stop = step => new Stop(step);
 
-module.exports = { goTo, branch, stop, journey, RequestBoundJourney };
+module.exports = {
+  goTo,
+  branch,
+  stop,
+  journey,
+  RequestBoundJourney,
+  ifCompleteThenContinue,
+  continueToNext,
+  stopHere,
+  validateThenStopHere
+};
