+++ src/steps/EntryPoint.js
@@ -1,5 +1,6 @@
 const Redirect = require('./Redirect');
 const createSession = require('../session/createSession');
+const { continueToNext } = require('../flow');
 
 class EntryPoint extends Redirect {
   get middleware() {
@@ -9,6 +10,10 @@ class EntryPoint extends Redirect {
   handler(req, res) {
     req.session.entryPoint = this.name;
     super.handler(req, res);
+  }
+
+  get flowControl() {
+    return continueToNext(this);
   }
 }
